Change sha256 test logic

+++ prepare-build.sh
@@ -1,11 +1,5 @@
 #!/usr/bin/env bash
 set -eu
-HASH_STATUS=0
-echo "$(cat PTD.apk.sha256) PTD.apk" | sha256sum --check --status || HASH_STATUS=$?
-if [ "$HASH_STATUS" = 1 ]; then
-  echo 'error: SHA256 hash did not match the expected value'
-  exit 1
-fi
 
 rm -rf ./out/cpp
 Il2CppInspector -i PTD.apk --select-outputs --cpp-out ./out/cpp
@@ -14,6 +8,13 @@ rm ./out/cpp/appdata/il2cpp-functions.h
 cp ./out/cpp/appdata/*.h ./hook/appdata/
 
 apktool d -f ./PTD.apk -o ./out/apktool
+APKTOOL_OUTPUT_EXPECTED_HASH="$(cat apktool-output-hash.txt)"
+APKTOOL_OUTPUT_HASH="$(nix hash-path ./out/apktool)"
+if [ "$APKTOOL_OUTPUT_EXPECTED_HASH" != "$APKTOOL_OUTPUT_HASH" ]; then
+  echo 'error: SHA256 hash did not match the expected value'
+  exit 1
+fi
+
 rm -rf ./out/main
 mkdir -p ./out/main
 cp ./out/apktool/lib/armeabi-v7a/libmain.so ./out/main/
